Renaming, remove results, README
