+++ parse/lexer.py
@@ -1,4 +1,4 @@
-# Generated from antlrparser/MySqlLexer.g4 by ANTLR 4.7.1
+# Generated from parse/MySqlLexer.g4 by ANTLR 4.7.1
 from antlr4 import *
 from io import StringIO
 from typing.io import TextIO
